efi: Convert RS pointers on SetVirtualAddressMap
When SetVirtualAddressMap called, the RuntimeServices should be
relocated from physical address to given virtual address. This requires
our firmware to be relocatable. It's still hard to implement with
current design. In this commit, it just convert the function pointers to
not_available() function that returns EFI_UNSUPPORTED to avoid
unexpected crash.

Signed-off-by: Akira Moroo <retrage01@gmail.com>