efi: Report text section as RuntimeServicesCode
Reporting the text section as RuntimeServicesData region causes
unhandled page fault in runtime phase on Ubuntu Focal. This is because
RuntimeServicesData region page is marked as non-executable. This commit
fixes this issue by reporting the section as RuntimeServicesCode. Since
EFI memory map is page-glanular, the text section must be page aligned.

Signed-off-by: Akira Moroo <retrage01@gmail.com>