build: Add initial support for aarch64 build
This commit adds initial support for aarch64 build. The goal is to boot
Linux with this firmware on aarch64 Cloud Hypervisor. This commit
includes a bootstrap assembly and build configs for building a binary
that complies the Linux kernel image format [1]. It also includes some
stub code to make the aarch64 target buildable with minimal changes.
The later commits will implement the code.

The current `pe::tests::test_loader` test uses an x86_64 specific disk
image. We need to provide aarch64 specific target for the test.

[1] https://docs.kernel.org/arm64/booting.html#call-the-kernel-image

Signed-off-by: Akira Moroo <retrage01@gmail.com>